feat:  define custom class and enum
Custom class establishes the `TextNode` type to handle the full text of
a line, the `TextType` expected, and optionally the URL that any link or
image formatting might contain.

Custom Enum `TextType` stores the general formatting string for the
following MD inline formatting:

- Normal Text
- Italic Text
- Bold Text
- Inline Code Text
- Links
- Image Links

I'm fairly certain that I wrote the Enum definitions wrong, so those are
entirely likely to change later on.