test: add test cases for new node type
Added a few test cases for the HTML test type.  There are some issues
with both the tests and the code that they are testing, but I think the
core of what they are testing is accurate.  Biggest issue is with the
`__repr__` for the HTMLNode type.  Need to make that a bit more dynamic
to allow for the appropriate type of quotations (`"`/`'`).