refactor: further distinguish between inlineutils
Really dislike doing multiple refactors one after the other, but at this
point it seems more logical to divide utilities this way at this point.
Added benefit, it leaves little guess work in which file to import for a
specific function since all of the functions share the same prefix to
the file's `-utils` name.