fix: explicit call to main
Didn't realize that I had omitted this line until now, and it hasn't
caused any issues up to this point that I'm aware of.  Regardless, I
figure it would be a good idea to ensure that my test scripts were all
written uniformly.

Is this a `fix`, or a `refactor`?