fix: actually use previously made functions
After beating my head against the keyboard enough, I realized that I had
previously made a function that would take markdown and split it into
individual blocks...so there was absolutely _no need_ for me to reinvent
the wheel!  Guess that's a testament to trying to knock out more of
these individual lessons in succession as opposed to "when the mood
strikes".